Users/hypark5540/downloads/django_react/requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
Django==2.2.5
django-cors-headers==3.1.0
django-jquery==3.1.0
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
lxml==4.3.4
mccabe==0.6.1
numpy==1.16.4
pandas==0.24.2
pep8==1.7.1
presto-python-client==0.7.0
pycodestyle==2.5.0
pylint==2.3.1
PyMySQL==0.9.3
python-dateutil==2.8.0
pytrends==4.6.0
pytz==2018.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.7.5
wrapt==1.11.2
